Finish up a tentative version
