+++ marketing_activity_extension/marketing-activities-template.js
